+++ ClientFactory/DummyClient.php
@@ -2,6 +2,8 @@
 
 namespace Http\HttplugBundle\ClientFactory;
 
+@trigger_error('The '.__NAMESPACE__.'\DummyClient interface is deprecated since version 1.7 and will be removed in 2.0.', E_USER_DEPRECATED);
+
 use Http\Client\HttpAsyncClient;
 use Http\Client\HttpClient;